Updated force update detction
Changed the 'Force Windows update detection cycle' resource from an execute to a powershell_script as it was failing with -2147024891(access denied) which I'm guessing is due to the execute resource running in ruby instead of a native shell. Also removed the c:\windows\sysnative prefix, as it should get resolved properly when ran through a Powershell shell.