Add classes to run WUA jobs with progress tracking
Implement a synchronous job system based on asynchronous actions to
allow progress tracking.

Asynchronous API normally use callbacks, but they do not seem to be
compatible with the ruby WIN32OLE implementation.
A small hack consist to pass a valid WIN32OLE object as dummy callback
pointer to be able to call BeginInstall & BeginDownload methods.

These classes will be used in the next commit to refactor the update
resource.